battery-status/plugging-1.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Test: battery not full, charger plugging in</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/battery-status/">
<meta name="flags" content="interact">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
  #note {
    background-color: #fef1b5;
    border: solid 1px #cdab2d;
    padding: 5px;
    margin: 15px;
    display: block;
  }
</style>

<h2>Description</h2>
<p>
  This test validates that all of the BatteryManager attributes exist and
  are set to correct values, with corresponding events fired,
  when the charger is plugged in.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    The device is unplugged from the charger before this test is run.
  </li>
  <li>
    The battery must not be full or reach full capacity before the time the test is run.
  </li>
</ol>

<div id="note">
  <ol>
    <li>
      Plug in the charger and wait for all the tests to complete.
    </li>
    <li>
      The tests may take long time since the definition of how
      often the chargingtimechange, dischargingtimechange, and levelchange
      events are fired is left to the implementation.
    </li>
  </ol>
</div>

<div id="log"></div>

<script>

(function() {

  setup({ explicit_timeout: true });

  var onchargingchange_test = async_test(
      'When the device is plugged in and its charging state is updated, ' +
      'must set the charging attribute\'s value to true and ' +
      'fire a chargingchange event.');

  var batterySuccess = function (battery) {
    battery.onchargingchange = onchargingchange_test.step_func(function () {
      assert_true(battery.charging, 'The charging attribute must be set to true.');
      onchargingchange_test.done();
    });
  };

  var batteryFailure = function (error) {
    onchargingchange_test.step(function () {
      assert_unreached(error.message);
    });
    onchargingchange_test.done();
  };

  navigator.getBattery().then(batterySuccess, batteryFailure);

})();

</script>
